+++ bgmi/lib/script.py
@@ -6,13 +6,12 @@ import imp
 import os
 import time
 import traceback
-import datetime
 
 from bgmi.config import SCRIPT_PATH, MAX_PAGE
-from bgmi.download import download_prepare
-from bgmi.models import STATUS_UPDATED, STATUS_FOLLOWED
-from bgmi.models import Scripts
-from bgmi.fetch import DATA_SOURCE_MAP
+from bgmi.lib.download import download_prepare
+from bgmi.lib.models import STATUS_UPDATED, STATUS_FOLLOWED
+from bgmi.lib.models import Scripts
+from bgmi.lib.fetch import DATA_SOURCE_MAP
 from bgmi.utils import print_success, print_warning, print_info
