+++ tests/test_controllers.py
@@ -4,9 +4,9 @@ from __future__ import print_function, unicode_literals
 import os
 import unittest
 
-from bgmi.controllers import *
+from bgmi.lib.controllers import *
 from bgmi.main import setup
-from bgmi.constants import unicode_
+from bgmi.lib.constants import unicode_
 
 
 class ControllersTest(unittest.TestCase):
